+++ database/migrations/2020_08_26_172148_create_periodos_table.php
@@ -20,7 +20,7 @@ class CreatePeriodosTable extends Migration
             $table->date('fecha_inicio');
             $table->date('fecha_fin');
 
-            $table->integer('grados_id')->unsigned();
+            $table->unsignedBigInteger('grados_id')->unsigned();
             $table->foreign('grados_id')->references('id')->on('grados')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
