+++ database/migrations/2020_08_27_013522_create_matriculas_table.php
@@ -22,7 +22,7 @@ class CreateMatriculasTable extends Migration
             $table->foreign('users_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
 
 
-            $table->integer('estudiantes_id')->unsigned();
+            $table->unsignedBigInteger('estudiantes_id')->unsigned();
             $table->foreign('estudiantes_id')->references('id')->on('estudiantes')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
